+++ test/libsolidity/smtCheckerTests/inline_assembly/empty.sol
@@ -8,5 +8,5 @@ contract C
 	}
 }
 // ----
-// Warning 7737: (76-90): Assertion checker does not support inline assembly.
-// Warning 7737: (76-90): Assertion checker does not support inline assembly.
+// Warning 7737: (76-90): Inline assembly may cause SMTChecker to produce spurious warnings (false positives).
+// Warning 7737: (76-90): Inline assembly may cause SMTChecker to produce spurious warnings (false positives).